requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0

# AI/ML
google-genai>=1.0.0
openai>=1.0.0

# Content extraction
readability-lxml>=0.8.1

# PDF processing (optional - for USP monographs)
PyMuPDF>=1.23.0

# Date parsing
python-dateutil>=2.8.0

# Email
# (uses built-in smtplib)
